+++ tracing/src/core/trace_writer_impl.h
@@ -38,9 +38,7 @@ class TraceWriterImpl : public TraceWriter,
   using TracePacket = pbzero::TracePacket;
   using TracePacketHandle = protozero::ProtoZeroMessageHandle<TracePacket>;
 
-  TraceWriterImpl(ProducerSharedMemoryArbiter*,
-                  WriterID,
-                  uint32_t target_buffer);
+  TraceWriterImpl(ProducerSharedMemoryArbiter*, WriterID, size_t target_buffer);
   ~TraceWriterImpl() override;
 
   // TraceWriter implementation.
@@ -67,7 +65,7 @@ class TraceWriterImpl : public TraceWriter,
 
   // This is just propagated into the chunk heaer.
   // See comments in DataSourceConfig::target_buffer.
-  const uint32_t target_buffer_;
+  const size_t target_buffer_;
 
   // Monotonic sequence id of the chunk. Together with the WriterID  is allows
   // the Service to to reconstruct the linear sequence of packets.
